@@ -19,18 +19,18 @@
     }
   ],
   "require": {
-    "php": "^7.3",
-    "cartalyst/tags": ">=v8.0.0",
+    "php": "^8.2",
+    "cartalyst/tags": "^14.0",
     "guzzlehttp/guzzle": "^7.0",
-    "illuminate/database": "^8.0",
-    "illuminate/support": "^8.0",
-    "illuminate/console": "^8.0",
-    "illuminate/contracts": "^8.0",
-    "illuminate/events": "^8.0",
-    "league/fractal": "^0.19.0"
+    "illuminate/database": "^11.0",
+    "illuminate/support": "^11.0",
+    "illuminate/console": "^11.0",
+    "illuminate/contracts": "^11.0",
+    "illuminate/events": "^11.0",
+    "league/fractal": "^0.20.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9.3"
+    "phpunit/phpunit": "^11.0.1"
   },
   "autoload": {
     "psr-4": {
