@@ -6,29 +6,29 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateCategoryTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('post_categories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('wp_id')->unique();
-            $table->string('name');
-            $table->string('slug');
-            $table->timestamps();
-        });
-    }
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::create('post_categories', function (Blueprint $table) {
+			$table->increments('id');
+			$table->integer('wp_id')->unique();
+			$table->string('name');
+			$table->string('slug');
+			$table->timestamps();
+		});
+	}
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('post_categories');
-    }
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+		Schema::dropIfExists('post_categories');
+	}
 }
