@@ -12,24 +12,24 @@ use League\Fractal\TransformerAbstract;
 
 class CategoryTransformer extends TransformerAbstract
 {
-    public function transform($post)
-    {
-        $embedded = collect($post->_embedded ?? []);
+	public function transform($post)
+	{
+		$embedded = collect($post->_embedded ?? []);
 
-        if ($embedded->has('wp:term')) {
+		if ($embedded->has('wp:term')) {
 
-            $category = $embedded->only('wp:term')
-                                 ->flatten(2)
-                                 ->where('taxonomy', 'category')
-                                 ->first();
+			$category = $embedded->only('wp:term')
+				->flatten(2)
+				->where('taxonomy', 'category')
+				->first();
 
-            return [
-                'wp_id' => $category->id,
-                'name'  => $category->name,
-                'slug'  => $category->slug,
-            ];
-        }
+			return [
+				'wp_id' => $category->id,
+				'name'  => $category->name,
+				'slug'  => $category->slug,
+			];
+		}
 
-        return [];
-    }
+		return [];
+	}
 }
